requirements.txt
Flask==2.3.3
pywebview==4.4.1
pyinstaller==6.3.0
